+++ src/app/within-isu-outcomes/within-isu-outcomes.component.html
@@ -5,7 +5,7 @@
   <h5>Would you like to add another?</h5>
 </div>
 <div *ngIf="!firstOutcome() && !nextOutcome()">
-  <h5>You cannot have more than {{max}} outcomes.</h5>
+  <h5>Whoa! That's a lot of outcomes!! We'll handle this lot, but if you really, really want to analyse more than {{max}} outcomes, consult a statstician... and good luck...</h5>
 </div>
 <form [formGroup]="outcomesForm">
   <div class="container-fluid">
